Moving the schema file
